server/requirements.txt
flask
Flask-PyMongo
flask_mail
flask_login
PyJWT
flask-socketio
python-dotenv
numpy
pandas
scikit-learn

torch
